backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
supabase==1.0.3
httpx>=0.23.0,<0.24.0
pillow==10.1.0
python-dotenv==1.0.0
transformers==4.35.0
torch==2.1.0
numpy==1.26.1
requests==2.31.0
